restructure visualisation
