+++ skactiveml/utils/__init__.py
@@ -10,8 +10,7 @@ from ._multi_annot import ext_confusion_matrix
 from ._selection import rand_argmax, rand_argmin
 from ._validation import check_classes, check_missing_label, check_scalar, \
     check_cost_matrix, check_classifier_params, check_X_y, MISSING_LABEL, \
-    check_random_state, check_class_prior, check_type
-from ._visualisation import mesh, check_bound
+    check_random_state, check_class_prior, check_type, check_bound
 
 __all__ = ['rand_argmax', 'rand_argmin', 'compute_vote_vectors',
            'majority_vote', 'is_unlabeled', 'is_labeled', 'ExtLabelEncoder',
@@ -19,5 +18,4 @@ __all__ = ['rand_argmax', 'rand_argmin', 'compute_vote_vectors',
            'check_scalar', 'check_classifier_params', 'check_X_y',
            'check_random_state', 'MISSING_LABEL', 'call_func', 'simple_batch',
            'check_class_prior', 'ext_confusion_matrix', 'fit_if_not_fitted',
-           'labeled_indices', 'unlabeled_indices', 'check_type', 'check_bound',
-           'mesh']
+           'labeled_indices', 'unlabeled_indices', 'check_type', 'check_bound']
